Moved WooPosCard.kt to components
